make the InferTypes.unify case more explicit
Problem: when a newcomer contributes to the project
for the first time, the comment in Elm/Data/Type.elm asking for adding
a case in the InfertTypes.unify function is meaningless
since the developer does not understand what this comment
is talking about. Thus, the developer could skip this comment,
leading to hard-to-find bug instead of easily understandable
compiler error message.

Indeed, the unify function pattern matchs against the Joker pattern "_"
at the end, silently ignoring new types.

Solution: Remove Joker pattern and pattern match against
    (T a b, T c d)

AND
    (T _ _, _)
for all types T.

There still is a comment, but this one is more local than the previous one.
Moreover this comment just asks "please do the same" instead of
"please do a specific thing".