simplify InferTypes/AssignIds.elm
Problem: the `AssignIds.elm` use the `IdSource` abstraction. This abstraction
garantees that all ids are different by only exposing functions that increment
the id. However, the code using this abstraction needs to carefuly deal with
the several `idSource` states, defeating the initial goal. Hence this abstraction
does not guarantee for what it was designed for, but add useless complexity
to the project.

Solution: remove the `IdSource` abstraction and use a bare `Int` instead
(aliased as `Id`).