Race condition fix? Who knows
https://travis-ci.com/github/elm-in-elm/compiler/builds/174354132

Travis recently failed because it logged "Compilation finished" *after* the
`WRITING TO FS` message, instead of before. This commit hopes to make it
always write it before (by wrapping the `WRITING TO FS` message writing
inside a `setTimeout(0)`), but it's inherently async so I'm a bit afraid
of the ordering being unpredictable and our snapshots testing being too
fine-grained (perhaps we shouldn't care about it? If we should, we
should move it from JS to Elm to get that synchronicity and
reproducibility).