we do not need partial results
This was an extra layer we do not need. Expressions are never done until they have to be done due to a newline. Therefore, we only ever used the progress variant unless we are parsing the newline token.

Because we now parse newline tokens at a nice an high level of the parser we can deal with them nicely.